exams for practice in the minterm
This hobbit was a very well-to-do hobbit, and his name was Baggins. The Bagginses had
lived in the neighbourhood of The Hill for time out of mind, and people considered them
very respectable, not only because most of them were rich, but also because they never
had any adventures or did anything unexpected: you could tell what a Baggins would
say on any question without the bother of asking him. This is a story of how a Baggins
had an adventure, found himself doing and saying things altogether unexpected. He may
have lost the neighbours' respect, but he gained-well, you will see whether he gained
anything in the end.